Implemented redundant path consistency
We check two things for a set of chromosomes: if they start with the first city, if not, shift them & whether the second city is smaller than the city at the end, otherwise the array is inversed (except for the first city). The check now is called twice after population initialization and after loopdetection heuristic.